+++ src/simputils/config/components/strategies/__init__.py
@@ -1,0 +1,2 @@
+from .MergingStrategyFlat import MergingStrategyFlat
+from .MergingStrategyRecursive import MergingStrategyRecursive