+++ src/simputils/config/enums/MergingStrategiesEnum.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class MergingStrategiesEnum(str, Enum):
+
+	FLAT = "flat"
+	RECURSIVE = "recursive"